Refactored the tests helpers to its own file
