+++ libvirt/helpers_test.go
@@ -2,6 +2,7 @@ package libvirt
 
 import (
 	"fmt"
+	"log"
 	"os"
 	"reflect"
 	"strings"
@@ -14,6 +15,8 @@ import (
 	"github.com/libvirt/libvirt-go-xml"
 	"github.com/terraform-providers/terraform-provider-ignition/ignition"
 )
+
+// This file contain function helpers used for testsuite/testacc
 
 var testAccProviders map[string]terraform.ResourceProvider
 var testAccProvider *schema.Provider
@@ -45,6 +48,60 @@ func testAccPreCheck(t *testing.T) {
 func testAccEnabled() bool {
 	v := os.Getenv("TF_ACC")
 	return v == "1" || strings.ToLower(v) == "true"
+}
+
+// //////////////////////////////////////////////////////////////////
+// general
+// //////////////////////////////////////////////////////////////////
+
+// getResourceFromTerraformState get aresource by name
+// from terraform states produced during testacc
+// and return the resource
+func getResourceFromTerraformState(resourceName string, state *terraform.State) (*terraform.ResourceState, error) {
+	rs, ok := state.RootModule().Resources[resourceName]
+	if !ok {
+		return nil, fmt.Errorf("Not found: %s", resourceName)
+	}
+
+	if rs.Primary.ID == "" {
+		return nil, fmt.Errorf("No libvirt resource key ID is set")
+	}
+	return rs, nil
+}
+
+// ** resource specifics helpers **
+
+// getVolumeFromTerraformState lookup volume by name and return the libvirt volume from a terraform state
+func getVolumeFromTerraformState(name string, state *terraform.State, virConn libvirt.Connect) (*libvirt.StorageVol, error) {
+	rs, err := getResourceFromTerraformState(name, state)
+	if err != nil {
+		return nil, err
+	}
+
+	vol, err := virConn.LookupStorageVolByKey(rs.Primary.ID)
+	if err != nil {
+		return nil, err
+	}
+	log.Printf("[DEBUG]:The ID is %s", rs.Primary.ID)
+	return vol, nil
+}
+
+// helper used in network tests for retrieve xml network definition.
+func getNetworkDef(state *terraform.State, name string, virConn libvirt.Connect) (*libvirtxml.Network, error) {
+	var network *libvirt.Network
+	rs, err := getResourceFromTerraformState(name, state)
+	if err != nil {
+		return nil, err
+	}
+	network, err = virConn.LookupNetworkByUUIDString(rs.Primary.ID)
+	if err != nil {
+		return nil, err
+	}
+	networkDef, err := getXMLNetworkDefFromLibvirt(network)
+	if err != nil {
+		return nil, fmt.Errorf("Error reading libvirt network XML description: %s", err)
+	}
+	return &networkDef, nil
 }
 
 // //////////////////////////////////////////////////////////////////
